Fix: Primitive type inferred based on parameter default value
