+++ common/changes/@webdoc/parser/fix-inferred-param-types_2022-02-19-18-06.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@webdoc/parser",
+      "comment": "Add primitive type inference for untyped parameters with a default value (https://github.com/webdoc-labs/webdoc/issues/138)",
+      "type": "none"
+    }
+  ],
+  "packageName": "@webdoc/parser",
+  "email": "sukantk2002@outlook.com"
+}
